tuples, all primitives, refactor
